+++ src/main/java/org/programers/vouchermanagement/voucher/dto/VoucherCreationRequest.java
@@ -1,6 +1,6 @@
-package org.programers.vouchermanagement.dto;
+package org.programers.vouchermanagement.voucher.dto;
 
-import org.programers.vouchermanagement.domain.VoucherPolicy;
+import org.programers.vouchermanagement.voucher.domain.VoucherPolicy;
 
 public class VoucherCreationRequest {
