+++ src/main/java/org/programers/vouchermanagement/voucher/dto/VoucherResponse.java
@@ -1,7 +1,7 @@
-package org.programers.vouchermanagement.dto;
+package org.programers.vouchermanagement.voucher.dto;
 
-import org.programers.vouchermanagement.domain.Voucher;
-import org.programers.vouchermanagement.domain.VoucherPolicy;
+import org.programers.vouchermanagement.voucher.domain.Voucher;
+import org.programers.vouchermanagement.voucher.domain.VoucherPolicy;
 
 import java.util.UUID;
