+++ src/main/java/org/programers/vouchermanagement/voucher/exception/NoSuchVoucherException.java
@@ -1,4 +1,4 @@
-package org.programers.vouchermanagement.exception;
+package org.programers.vouchermanagement.voucher.exception;
 
 public class NoSuchVoucherException extends RuntimeException {
