+++ src/test/java/org/programers/vouchermanagement/voucher/domain/FixedAmountPolicyTest.java
@@ -1,6 +1,7 @@
-package org.programers.vouchermanagement.domain;
+package org.programers.vouchermanagement.voucher.domain;
 
 import org.junit.jupiter.api.Test;
+import org.programers.vouchermanagement.voucher.domain.FixedAmountPolicy;
 
 import static org.assertj.core.api.Assertions.assertThat;
