+++ src/test/java/org/programers/vouchermanagement/voucher/domain/PercentDiscountPolicyTest.java
@@ -1,6 +1,7 @@
-package org.programers.vouchermanagement.domain;
+package org.programers.vouchermanagement.voucher.domain;
 
 import org.junit.jupiter.api.Test;
+import org.programers.vouchermanagement.voucher.domain.PercentDiscountPolicy;
 
 import static org.assertj.core.api.Assertions.assertThat;
